Move WORKER_IDENTITY
The previous placement didn't work as the main infrahub module was
imported by gunicorn so all workers would get the same ID. 🤦‍♂️
