+++ backend/infrahub/worker.py
@@ -1,0 +1,3 @@
+import uuid
+
+WORKER_IDENTITY = str(uuid.uuid4())
